requirements.txt
streamlit
opencv-python
pillow
numpy
pytesseract
streamlit-webrtc
deep-translator
gtts
